Refactor telemetry behaviors, update mediator API
- Split TelemetryBehaviors.cs into four focused files for each telemetry behavior class.
- Add GenDI package reference and update project references in NetMediate.Diagnostics.csproj.
- Update IMediator extension generation: Notify methods are now void (with batch/keyed overloads), Send methods use ValueTask and batch methods are renamed to BatchAsync.
- Update CoreDispatchBenchmarks to use new async extension methods and ValueTask-based warmup.
- Refactor DrainStream for clarity.
