+++ src/NetMediate.Diagnostics/NetMediate.Diagnostics.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 	<Import Project="..\..\pack.props" />
 	<PropertyGroup>
 		<AssemblyName>NetMediate.Diagnostics</AssemblyName>
@@ -15,6 +15,12 @@
 		<None Include="..\..\LICENSE" Pack="true" PackagePath="\" />
 	</ItemGroup>
 	<ItemGroup>
-		<ProjectReference Include="..\NetMediate\NetMediate.csproj" />
+		<PackageReference Include="GenDI" Version="2026.5.19.1218"/>
+	</ItemGroup>
+	<ItemGroup>
+	  <ProjectReference Include="..\NetMediate.Core\NetMediate.Core.csproj" />
+	</ItemGroup>
+	<ItemGroup>
+		<Using Include="GenDI"/>
 	</ItemGroup>
 </Project>
